Сделал DTO для DogOwner и CatOwner, изменил контроллеры и сервисы.

+++ src/main/java/com/devsteam/getname/telbot_shelterdc/repository/DogOwnerRepository.java
@@ -1,10 +1,9 @@
 package com.devsteam.getname.telbot_shelterdc.repository;
 
-import com.devsteam.getname.telbot_shelterdc.model.Dog;
 import com.devsteam.getname.telbot_shelterdc.model.DogOwner;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DogOwnerRepository extends JpaRepository<DogOwner, Integer> {
+public interface DogOwnerRepository extends JpaRepository<DogOwner, Long> {
 }
